requirements.txt
requests
sentence-transformers
faiss-cpu
transformers
flask

